reorganized modelCatalog Routes
