+++ frontend/src/routes/modelCatalog/detailsRoutes.ts
