+++ frontend/src/routes/modelCatalog/index.ts
@@ -1,2 +1,2 @@
-export * from './details/routes';
-export * from './register/routes';
+export * from './detailsRoutes';
+export * from './registerRoutes';
