+++ frontend/src/routes/modelCatalog/registerRoutes.ts
